backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
requests
python-dotenv
starlette
numpy